requirements.txt
Flask
python-dotenv
google-generativeai
werkzeug
gunicorn
pymongo[srv]
dnspython
